+++ examples/Example_17.java
@@ -247,7 +247,7 @@ public class Example_17 {
         fis = new FileInputStream(fileName);
         Image image54 = new Image(pdf, fis);
 
-        fileName = "images/LGK_ADDRESS.PNG";
+        fileName = "images/rgba-8bit-chunks.png";
         fis = new FileInputStream(fileName);
         Image image55 = new Image(pdf, fis);