+++ examples/Example_51/Example_51.cs
@@ -23,7 +23,7 @@ public class Example_51 {
     public Example_51() {
         List<PDFobj> objects;
         using (BufferedStream stream = new BufferedStream(
-                new FileStream("data/testPDFs/wirth.pdf", FileMode.Open, FileAccess.Read))) {
+                new FileStream("data/testPDFs/scanned-linearized.pdf", FileMode.Open, FileAccess.Read))) {
             objects = new PDF().Read(stream);
         }
         int count = new PDF().GetPageObjects(objects).Count;